Game Set Up - Major Progress
1. Created two new classes GameData and LevelStruct
   - GameData handles loading games / saving games
   - LevelStruct handles the actual events inside of levels

2. Moved gameSetUp() to GameData class

3. Updated Makefile

4. Rewrote structure and most sections of ReadMe to reflect
 current state of project better
