+++ basefile.h
@@ -3,12 +3,15 @@
 #define BASEFILE_H
 
 #include <iostream>
+#include <thread>
 
 
 #include "items/Item.h"
 #include "items/InventoryList.h"
 #include "items/derivedItems/Healing.h"
 #include "entities/Entity.h"
+#include "GameData.h"
+#include "LevelStruct.h"
 
 
 #endif
